+++ src/components/wrappers/MainLayout.tsx
@@ -2,7 +2,7 @@ import { Drawer } from "@material-ui/core";
 import React from "react";
 import { RouteComponentProps } from "react-router-dom";
 import Navigator from "src/components/molecules/Navigator";
-import { NotificationListenerProps } from "src/contexts/NotificationListener";
+import { NotificationListenerProps } from "src/components/wrappers/NotificationListener";
 import Locale, { location } from "src/localization/locale";
 import styled from "styled-components";
